+++ tests/testthat/test-interrupt.R
@@ -1,11 +1,6 @@
 test_that("keypress() is interruptible", {
-  skip_on_os("windows")
   skip_on_cran()
 
-  # Run keypress() in a child R process, attached to a pseudo-terminal so
-  # that has_keypress_support() is satisfied and keypress() actually blocks.
-  # We then send a SIGINT and check that the blocking read is interrupted
-  # cleanly, i.e. the interrupt condition is raised and can be caught.
   opts <- callr::r_process_options(
     func = function() {
       # Make sure has_keypress_support() is happy in the child.
@@ -22,14 +17,19 @@ test_that("keypress() is interruptible", {
   p <- callr::r_process$new(opts)
   on.exit(p$kill(), add = TRUE)
 
-  # Wait until the child signals that it is about to block in keypress().
+  # on Windows R adds a "set title" ANSI sequence
+  strip_ansi <- function(x) {
+    gsub("\033(?:\\[[0-9;?]*[A-Za-z]|\\][^\007]*\007)", "", x, perl = TRUE)
+  }
   out <- ""
   deadline <- Sys.time() + 10
-  while (Sys.time() < deadline && !grepl("READY", out, fixed = TRUE)) {
+  while (
+    Sys.time() < deadline && !grepl("READY", strip_ansi(out), fixed = TRUE)
+  ) {
     p$poll_io(200)
     out <- paste0(out, p$read_output())
   }
-  expect_match(out, "READY", fixed = TRUE)
+  expect_match(strip_ansi(out), "READY", fixed = TRUE)
 
   # Give keypress() a moment to enter its poll loop, then interrupt it.
   Sys.sleep(0.5)